[#44] move PassWordEncoders
