+++ src/main/java/com/safecornerscoffee/borders/security/password/PasswordEncoder.java
@@ -1,4 +1,4 @@
-package com.safecornerscoffee.borders.helper;
+package com.safecornerscoffee.borders.security.password;
 
 public interface PasswordEncoder {
