+++ src/test/java/com/safecornerscoffee/borders/helper/BcryptPasswordEncoderTest.java
@@ -1,5 +1,7 @@
 package com.safecornerscoffee.borders.helper;
 
+import com.safecornerscoffee.borders.security.password.BcryptPasswordEncoder;
+import com.safecornerscoffee.borders.security.password.PasswordEncoder;
 import org.junit.Test;
 
 import static org.assertj.core.api.Assertions.*;
